Create login form class and template
The login template is now rendered when visiting login/ and the
appropriate login ticket ID is generated, stored in the DB and
displayed as a hidden field on the form. We may want to eventually
replace this with the built-in Django CSRF functionality, but for
now we'll implement the CAS spec as-is.