validate_pgturl: Respecting REQUESTS_CA_BUNDLE environment variable.
This allows, for example, a local development environment with a self-signed
cert installed as its root CA, and the following:

    %sudo update-ca-certificates
    %export REQUESTS_CA_BUNDLE='/etc/ssl/certs/ca-certificates.crt'

The certificate can then be accepted as valid.

While mama_cas-specific settings are expected to live in settings.py, this env
var is the standard way to specify this value for requests.